backend/requirements.txt
# TODO Pick package versions to use [#3]
boto3
django
django-cors-headers
django-import-export
django-storages
djangorestframework
psycopg[binary,pool]
